pyproject.toml
[tool.poetry]
packages = [{include = "papl", from = "."}]
name = "papl"
version = "0.1.0"
description = ""
authors = ["Nicolas Durrande"]
license = "Apache 2.0"

[tool.poetry.dependencies]
python = "^3.8"
gpflow = "^2.2.1"
gpflux = "^0.2.3"
matplotlib = "^3.3.3"
jupyterlab = "^3.3.1"
tensorflow-probability = "<=0.13"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.dev-dependencies]
taskipy = "^1.2.0"
black = "^22.1.0"
isort = "^5.10.1"
autoflake = "^1.4"
flake8 = "^4.0.1"
pytest = "^7.0.1"

[tool.taskipy.tasks]
format = "autoflake --remove-all-unused-imports --recursive --remove-unused-variables --in-place --exclude=__init__.py papl tests && black papl tests && isort papl tests"
check = "flake8 --statistics --exclude=.ipynb_checkpoints tests papl && black --check --diff tests papl && isort --check --diff tests papl"
test = "pytest -v -x --ff -rN -Wignore -s --tb=short --durations=10"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black" 
